Complete Fix bug, validation, UI and permission

+++ app/Controllers/API/Bankquestion.php
@@ -27,12 +27,16 @@ class Bankquestion extends BaseController
 
    public function create()
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+
       $validation = \Config\Services::validation();
       $validation->setRules($this->rules, $this->errors);
       parse_str(file_get_contents('php://input'), $input);
       if ($validation->withRequest($this->request)->run() == false) {
          return $this->respond([
-            "message" => "Failed to save changes.",
+            "message" => "Failed to added.",
             "status" => 400,
             "errors" => $validation->getErrors()
          ]);
@@ -59,6 +63,16 @@ class Bankquestion extends BaseController
 
    public function update($bank_question_id)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
+      if ($this->role == 'teacher') {
+         if (!$this->m_bank_question->have_bank_question($this->username, $bank_question_id)) {
+            return $this->failForbidden();
+         }
+      }
+
       $validation = \Config\Services::validation();
       $validation->setRules($this->rules, $this->errors);
       parse_str(file_get_contents('php://input'), $input);
@@ -93,6 +107,16 @@ class Bankquestion extends BaseController
 
    public function delete($bank_question_id)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
+      if ($this->role == 'teacher') {
+         if (!$this->m_bank_question->have_bank_question($this->username, $bank_question_id)) {
+            return $this->failForbidden();
+         }
+      }
+
       $where = [
          "bank_question_id" => $bank_question_id
       ];
@@ -114,18 +138,22 @@ class Bankquestion extends BaseController
 
    public function show_question($bank_question_id, $number_question)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
       $result = $this->m_bank_question->get_question($bank_question_id, $number_question);
       if ($result) {
          $result->choices = json_decode($result->choice);
          return $this->respond([
-            "message" => "Data found!",
+            "message" => "Question found!",
             "status" => 200,
             "data" => $result,
             "error" => false
          ]);
       }
       return $this->respond([
-         "message" => "Data not found!",
+         "message" => "Question not found!",
          "status" => 200,
          "data" => null,
          "error" => true
@@ -134,6 +162,16 @@ class Bankquestion extends BaseController
 
    public function add_question($bank_question_id)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
+      if ($this->role == 'teacher') {
+         if (!$this->m_bank_question->have_bank_question($this->username, $bank_question_id)) {
+            return $this->failForbidden();
+         }
+      }
+      
       parse_str(file_get_contents('php://input'), $input);
       $question_ids = isset($input['question_id']) ? $input['question_id'] : [];
       $question_ids = array_map(function ($v) {
@@ -141,19 +179,19 @@ class Bankquestion extends BaseController
       }, $question_ids);
       $questions = json_decode($this->m_bank_question->questions($bank_question_id));
       $questions = array_merge($questions, $question_ids);
-      $questions = array_reverse($questions);
-      $questions = array_unique($questions);
-      $questions = array_reverse($questions);
+      $questions = array_reverse($questions); // reverse question
+      $questions = array_unique($questions); // remove duplicate question
+      $questions = array_reverse($questions); // reverse back
       $result = $this->m_bank_question->update_question($bank_question_id, $questions);
       if ($result) {
          return $this->respond([
-            "message" => "Changes saved successfully.",
+            "message" => "Added question successfully.",
             "status" => 200,
             "error" => false
          ]);
       }
       return $this->respond([
-         "message" => "Failed to save changes.",
+         "message" => "Failed to added.",
          "status" => 400,
          "error" => true
       ]);
@@ -161,6 +199,16 @@ class Bankquestion extends BaseController
 
    public function create_question($bank_question_id)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
+      if ($this->role == 'teacher') {
+         if (!$this->m_bank_question->have_bank_question($this->username, $bank_question_id)) {
+            return $this->failForbidden();
+         }
+      }
+
       parse_str(file_get_contents('php://input'), $input);
       if ($errors = (new Question)->question_validation($input, $this->request)) {
          return $this->respond([
@@ -177,7 +225,7 @@ class Bankquestion extends BaseController
             $choices[] = htmlentities($value, ENT_QUOTES, 'UTF-8');
          }
       }
-      $answer_key = $input['answer_key'] ? htmlentities($input['answer_key'], ENT_QUOTES, 'UTF-8') : null;
+      $answer_key = empty($input['answer_key']) ? null : htmlentities($input['answer_key'], ENT_QUOTES, 'UTF-8');
       $data = [
          "question_type" => $question_type,
          "question_text" => $question_text,
@@ -191,7 +239,7 @@ class Bankquestion extends BaseController
          $result = $this->m_bank_question->update_question($bank_question_id, $question_id);
          if ($result) {
             return $this->respond([
-               "message" => "Added successfully.",
+               "message" => "Added question successfully.",
                "status" => 200,
                "error" => false
             ]);
@@ -206,6 +254,16 @@ class Bankquestion extends BaseController
 
    public function delete_question($bank_question_id, $number_question)
    {
+      if ($this->role == 'student') {
+         return $this->failForbidden();
+      }
+      
+      if ($this->role == 'teacher') {
+         if (!$this->m_bank_question->have_bank_question($this->username, $bank_question_id)) {
+            return $this->failForbidden();
+         }
+      }
+
       try {
          $this->m_bank_question->delete_question($bank_question_id, $number_question);
          return $this->respond([
